backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-jwt-extended==4.7.1
flask-sqlalchemy==3.1.1
werkzeug==3.1.3
anthropic
Pillow
